Module vectis_contract_tests::common::common

Modules

sha256

SHA256

Macros

anyhow

Construct an ad-hoc error from a string or existing non-anyhow error value.

Structs

Addr

A human readable address.

App

Router is a persisted state. You can query this. Execution generally happens on the RouterCache, which then can be atomically committed or rolled back. We offer .execute() as a wrapper around cache, execute, commit/rollback process.

AppResponse

BalanceResponse

Binary

Binary is a wrapper around Vec to add base64 de/serialization with serde. It also adds some helper methods to help encode inline.

BlockInfo

CanonicalAddr

A blockchain address in its binary form.

Coin

ConfigResponse

ContractWrapper

Wraps the exported functions from a contract and provides the normalized format Place T4 and E4 at the end, as we just want default placeholders for most contracts that don’t have sudo

CreateWalletMsg

Empty

An empty struct that serves as a placeholder in different places, such as contracts that don’t set a custom message.

Event

A full Cosmos SDK event.

Guardians

GuardiansUpdateMsg

If the Guardians.guardian_multisig is given, we will instantiate a new multisig contract. This contract can be an instance of 2 code ids.

GuardiansUpdateRequest

InstantiateMsg

Member

A group member has a weight associated with them. This may all be equal, or may have meaning in the app that makes use of the group (eg. voting power)

Message

A (hashed) message input to an ECDSA signature

MultiSig

PRegistryInstantiateMsg

Plugin

PluginListResponse

PluginParams

PluginRegistry

PluginsResponse

PublicKey

A Secp256k1 public key, used for verification of signatures

RelayTransaction

Secp256k1

The secp256k1 engine, used to execute all signature operations

SecretKey

Secret 256-bit key used as x in an ECDSA signature

Uint128

A thin wrapper around u128 that is using strings for JSON encoding/decoding, such that the full u128 range can be used for clients that convert JSON numbers to floats, like JavaScript and jq.

VoterListResponse

cw3flexInstMsg

cw4InstMsg

Enums

CodeIdType

CosmosMsg

Duration

Duration is a delta of time. You can add it to a BlockInfo or Expiration to move that further in the future. Note that an height-based Duration and a time-based Expiration cannot be combined

Expiration

Expiration represents a point in time when some event happens. It can compare with a BlockInfo and will return is_expired() == true once the condition is hit (and for every block in the future)

FactoryExecuteMsg

FactoryQueryMsg

MultiSigQueryMsg

PRegistryContractError

PRegistryExecMsg

PRegistryQueryMsg

ProxyExecuteMsg

ProxyQueryMsg

QueryRequest

StdError

Structured error type for init, execute and query.

VectisActors

WalletFactoryExecuteMsg

WasmMsg

The message types of the wasm module.

WasmQuery

cw3flexExecMsg

cw4ExecMsg

Constants

CONTROLLER_ADDR

CONTROLLER_PRIV

DENOM

This is used for staking queries https://github.com/CosmWasm/cosmwasm/blob/32f308a1a56ae5b8278947891306f7a374c3df94/packages/vm/src/environment.rs#L383

GUARD1

GUARD2

GUARD3

INSTALL_FEE

MINTER_CAP

MINT_AMOUNT

MULTISIG_THRESHOLD

NON_CONTROLLER_PRIV

PROP_APPROVER

REGISTRY_FEE

WALLET_FEE

Traits

Contract

Interface to call into a Contract

DeserializeOwned

A data structure that can be deserialized without borrowing any data from the deserializer.

Executor

Serialize

A data structure that can be serialized into any data format supported by Serde.

Functions

coin

A shorthand constructor for Coin

contract_cw4

contract_factory

contract_fixed_multisig

contract_flex_multisig

contract_plugin_registry

contract_proxy

cw4_execute

cw4_instantiate

cw4_query

factory_execute

factory_instantiate

factory_query

factory_reply

fixed_multisig_execute

fixed_multisig_instantiate

fixed_multisig_query

flex_multisig_execute

flex_multisig_instantiate

flex_multisig_query

mock_dependencies

Creates all external requirements that can be injected for unit tests.

proxy_exec

proxy_execute

proxy_instantiate

proxy_migrate

proxy_query

proxy_reply

pub_key_to_address

Converts controller pubkey into Addr

to_binary

Type Definitions

Result

Result<T, Error>

StdResult

The return type for init, execute and query. Since the error type cannot be serialized to JSON, this is only available within the contract and its unit tests.

ThresholdAbsoluteCount

Declares that a fixed weight of Yes votes is needed to pass. See ThresholdResponse.AbsoluteCount in the cw3 spec for details. Only Fixed multisig is supported in this version

Derive Macros

Derivative

Serialize