Enum vectis_contract_tests::common::common::DaoExecMsg

pub enum DaoExecMsg {

    ExecuteAdminMsgs {
        msgs: Vec<CosmosMsg<Empty>, Global>,
    },
    ExecuteProposalHook {
        msgs: Vec<CosmosMsg<Empty>, Global>,
    },
    Pause {
        duration: Duration,
    },
    Receive(Cw20ReceiveMsg),
    ReceiveNft(Cw721ReceiveMsg),
    RemoveItem {
        key: String,
    },
    SetItem {
        key: String,
        value: String,
    },
    NominateAdmin {
        admin: Option<String>,
    },
    AcceptAdminNomination {},
    WithdrawAdminNomination {},
    UpdateConfig {
        config: Config,
    },
    UpdateCw20List {
        to_add: Vec<String, Global>,
        to_remove: Vec<String, Global>,
    },
    UpdateCw721List {
        to_add: Vec<String, Global>,
        to_remove: Vec<String, Global>,
    },
    UpdateProposalModules {
        to_add: Vec<ModuleInstantiateInfo, Global>,
        to_disable: Vec<String, Global>,
    },
    UpdateVotingModule {
        module: ModuleInstantiateInfo,
    },
    UpdateSubDaos {
        to_add: Vec<SubDao, Global>,
        to_remove: Vec<String, Global>,
    },

}

Variants

ExecuteAdminMsgs

Fields

msgs: Vec<CosmosMsg<Empty>, Global>

Callable by the Admin, if one is configured. Executes messages in order.

ExecuteProposalHook

Fields

msgs: Vec<CosmosMsg<Empty>, Global>

Callable by proposal modules. The DAO will execute the messages in the hook in order.

Pause

Fields

duration: Duration

Pauses the DAO for a set duration. When paused the DAO is unable to execute proposals

Receive(Cw20ReceiveMsg)

Executed when the contract receives a cw20 token. Depending on the contract’s configuration the contract will automatically add the token to its treasury.

ReceiveNft(Cw721ReceiveMsg)

Executed when the contract receives a cw721 token. Depending on the contract’s configuration the contract will automatically add the token to its treasury.

RemoveItem

Fields

key: String

Removes an item from the governance contract’s item map.

SetItem

Fields

key: String

value: String

Adds an item to the governance contract’s item map. If the item already exists the existing value is overriden. If the item does not exist a new item is added.

NominateAdmin

Fields

admin: Option<String>

Callable by the admin of the contract. If ADMIN is None the admin is set as the contract itself so that it may be updated later by vote. If ADMIN is Some a new admin is proposed and that new admin may become the admin by executing the AcceptAdminNomination message.

If there is already a pending admin nomination the WithdrawAdminNomination message must be executed before a new admin may be nominated.

AcceptAdminNomination

Fields

Callable by a nominated admin. Admins are nominated via the NominateAdmin message. Accepting a nomination will make the nominated address the new admin.

Requiring that the new admin accepts the nomination before becoming the admin protects against a typo causing the admin to change to an invalid address.

WithdrawAdminNomination

Fields

Callable by the current admin. Withdraws the current admin nomination.

UpdateConfig

Fields

config: Config

Callable by the core contract. Replaces the current governance contract config with the provided config.

UpdateCw20List

Fields

to_add: Vec<String, Global>

to_remove: Vec<String, Global>

Updates the list of cw20 tokens this contract has registered.

UpdateCw721List

Fields

to_add: Vec<String, Global>

to_remove: Vec<String, Global>

Updates the list of cw721 tokens this contract has registered.

UpdateProposalModules

Fields

to_add: Vec<ModuleInstantiateInfo, Global>

NOTE: the pre-propose-base package depends on it being the case that the core module instantiates its proposal module.

to_disable: Vec<String, Global>

Updates the governance contract’s governance modules. Module instantiate info in to_add is used to create new modules and install them.

UpdateVotingModule

Fields

module: ModuleInstantiateInfo

Callable by the core contract. Replaces the current voting module with a new one instantiated by the governance contract.

UpdateSubDaos

Fields

to_add: Vec<SubDao, Global>

to_remove: Vec<String, Global>

Update the core module to add/remove SubDAOs and their charters