Enum vectis_contract_tests::common::common::GovecExecuteMsg

pub enum GovecExecuteMsg {
    Transfer {
        recipient: String,
        amount: Uint128,
        relayed_from: Option<String>,
    },
    TransferFrom {
        owner: String,
        recipient: String,
        amount: Uint128,
    },
    Burn {
        amount: Uint128,
    },
    Exit {
        relayed_from: Option<String>,
    },
    Send {
        contract: String,
        amount: Uint128,
        msg: Binary,
        relayed_from: Option<String>,
    },
    Mint {
        new_wallet: String,
    },
    UpdateMintCap {
        new_mint_cap: Option<Uint128>,
    },
    UpdateMintAmount {
        new_amount: Uint128,
    },
    UpdateDaoAddr {
        new_addr: String,
    },
    UpdateMarketing {
        project: Option<String>,
        description: Option<String>,
        marketing: Option<String>,
    },
    UploadLogo(Logo),
}

Variants

Transfer

Fields

recipient: String

amount: Uint128

relayed_from: Option<String>

Transfer is a base message to move tokens to another account without triggering actions

TransferFrom

Fields

owner: String

recipient: String

amount: Uint128

Only the pre-proposal approval single contract can call this message and recieve tokens For Approval, Vectis plugins should be used

Burn

Fields

amount: Uint128

Burn amount specified from total supply permission: executed by dao only

Exit

Fields

relayed_from: Option<String>

Exits the DAO but removing itself from the ledger QueryMsg::Joined will return None

Send

Fields

contract: String

amount: Uint128

msg: Binary

relayed_from: Option<String>

Send is a base message to transfer tokens to a contract and trigger an action on the receiving contract.

Mint

Fields

new_wallet: String

If authorised, mint the MINT_AMOUNT defined by the dao to the new wallet permission: minters

UpdateMintCap

Fields

new_mint_cap: Option<Uint128>

Updates the mint cap of the contract.Authorized by the DAO permission: executed by dao only

UpdateMintAmount

Fields

new_amount: Uint128

Updates the mint cap of the contract.Authorized by the DAO permission: executed by dao only

UpdateDaoAddr

Fields

new_addr: String

Updates the staking contract address.Authorized by the DAO permission: executed by dao only

UpdateMarketing

Fields

project: Option<String>

A URL pointing to the project behind this token.

description: Option<String>

A longer description of the token and it’s utility. Designed for tooltips or such

marketing: Option<String>

The address (if any) who can update this data structure

If authorized, updates marketing metadata. Setting None/null for any of these will leave it unchanged. Setting Some(“”) will clear this field on the contract storage permission: executed by dao only

UploadLogo(Logo)

If set as the “marketing” role on the contract, upload a new URL, SVG, or PNG for the token