Enum vectis_contract_tests::common::common::PropQueryMsg

pub enum PropQueryMsg {

    Config {},
    Proposal {
        proposal_id: u64,
    },
    ListProposals {
        start_after: Option<u64>,
        limit: Option<u64>,
    },
    ReverseProposals {
        start_before: Option<u64>,
        limit: Option<u64>,
    },
    GetVote {
        proposal_id: u64,
        voter: String,
    },
    ListVotes {
        proposal_id: u64,
        start_after: Option<String>,
        limit: Option<u64>,
    },
    ProposalCount {},
    ProposalCreationPolicy {},
    ProposalHooks {},
    VoteHooks {},
    Dao {},
    Info {},
    NextProposalId {},

}

Variants

Config

Fields

Gets the proposal module’s config.

Proposal

Fields

proposal_id: u64

Gets information about a proposal.

ListProposals

Fields

start_after: Option<u64>

The proposal ID to start listing proposals after. For example, if this is set to 2 proposals with IDs 3 and higher will be returned.

limit: Option<u64>

The maximum number of proposals to return as part of this query. If no limit is set a max of 30 proposals will be returned.

Lists all the proposals that have been cast in this module.

ReverseProposals

Fields

start_before: Option<u64>

The proposal ID to start listing proposals before. For example, if this is set to 6 proposals with IDs 5 and lower will be returned.

limit: Option<u64>

The maximum number of proposals to return as part of this query. If no limit is set a max of 30 proposals will be returned.

Lists all of the proposals that have been cast in this module in decending order of proposal ID.

GetVote

Fields

proposal_id: u64

voter: String

Returns a voters position on a propsal.

ListVotes

Fields

proposal_id: u64

The proposal to list the votes of.

start_after: Option<String>

The voter to start listing votes after. Ordering is done alphabetically.

limit: Option<u64>

The maximum number of votes to return in response to this query. If no limit is specified a max of 30 are returned.

Lists all of the votes that have been cast on a proposal.

ProposalCount

Fields

Returns the number of proposals that have been created in this module.

ProposalCreationPolicy

Fields

Gets the current proposal creation policy for this module.

ProposalHooks

Fields

Lists all of the consumers of proposal hooks for this module.

VoteHooks

Fields

Lists all of the consumers of vote hooks for this module.

Dao

Fields

Returns the address of the DAO this module belongs to

Info

Fields

Returns contract version info

NextProposalId

Fields

Returns the proposal ID that will be assigned to the next proposal created.