Enum vectis_contract_tests::common::common::WasmMsg

#[non_exhaustive]
pub enum WasmMsg {
    Execute {
        contract_addr: String,
        msg: Binary,
        funds: Vec<Coin, Global>,
    },
    Instantiate {
        admin: Option<String>,
        code_id: u64,
        msg: Binary,
        funds: Vec<Coin, Global>,
        label: String,
    },
    Migrate {
        contract_addr: String,
        new_code_id: u64,
        msg: Binary,
    },
    UpdateAdmin {
        contract_addr: String,
        admin: String,
    },
    ClearAdmin {
        contract_addr: String,
    },
}

The message types of the wasm module.

See https://github.com/CosmWasm/wasmd/blob/v0.14.0/x/wasm/internal/types/tx.proto

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Execute

Fields

contract_addr: String

msg: Binary

msg is the json-encoded ExecuteMsg struct (as raw Binary)

funds: Vec<Coin, Global>

Dispatches a call to another contract at a known address (with known ABI).

This is translated to a MsgExecuteContract. sender is automatically filled with the current contract’s address.

Instantiate

Fields

admin: Option<String>

code_id: u64

msg: Binary

msg is the JSON-encoded InstantiateMsg struct (as raw Binary)

funds: Vec<Coin, Global>

label: String

A human-readbale label for the contract

Instantiates a new contracts from previously uploaded Wasm code.

The contract address is non-predictable. But it is guaranteed that when emitting the same Instantiate message multiple times, multiple instances on different addresses will be generated. See also Instantiate2.

This is translated to a MsgInstantiateContract. sender is automatically filled with the current contract’s address.

Migrate

Fields

contract_addr: String

new_code_id: u64

the code_id of the new logic to place in the given contract

msg: Binary

msg is the json-encoded MigrateMsg struct that will be passed to the new code

Migrates a given contracts to use new wasm code. Passes a MigrateMsg to allow us to customize behavior.

Only the contract admin (as defined in wasmd), if any, is able to make this call.

This is translated to a MsgMigrateContract. sender is automatically filled with the current contract’s address.

UpdateAdmin

Fields

contract_addr: String

admin: String

Sets a new admin (for migrate) on the given contract. Fails if this contract is not currently admin of the target contract.

ClearAdmin

Fields

contract_addr: String

Clears the admin on the given contract, so no more migration possible. Fails if this contract is not currently admin of the target contract.