Struct vectis_contract_tests::common::common::App

pub struct App<Bank = BankKeeper, Api = MockApi, Storage = MemoryStorage, Custom = FailingModule<Empty, Empty, Empty>, Wasm = WasmKeeper<Empty, Empty>, Staking = StakeKeeper, Distr = DistributionKeeper, Ibc = FailingModule<IbcMsg, IbcQuery, Empty>, Gov = FailingModule<GovMsg, Empty, Empty>> { /* private fields */ }

Router is a persisted state. You can query this. Execution generally happens on the RouterCache, which then can be atomically committed or rolled back. We offer .execute() as a wrapper around cache, execute, commit/rollback process.

Implementations

impl App<BankKeeper, MockApi, MemoryStorage, FailingModule<Empty, Empty, Empty>, WasmKeeper<Empty, Empty>, StakeKeeper, DistributionKeeper, FailingModule<IbcMsg, IbcQuery, Empty>, FailingModule<GovMsg, Empty, Empty>>

pub fn new<F>(
init_fn: F
) -> App<BankKeeper, MockApi, MemoryStorage, FailingModule<Empty, Empty, Empty>, WasmKeeper<Empty, Empty>, StakeKeeper, DistributionKeeper, FailingModule<IbcMsg, IbcQuery, Empty>, FailingModule<GovMsg, Empty, Empty>>where
F: FnOnce(&mut Router<BankKeeper, FailingModule<Empty, Empty, Empty>, WasmKeeper<Empty, Empty>, StakeKeeper, DistributionKeeper, FailingModule<IbcMsg, IbcQuery, Empty>, FailingModule<GovMsg, Empty, Empty>>, &dyn Api, &mut dyn Storage),

Creates new default App implementation working with Empty custom messages.

impl<BankT, ApiT, StorageT, CustomT, WasmT, StakingT, DistrT, IbcT, GovT> App<BankT, ApiT, StorageT, CustomT, WasmT, StakingT, DistrT, IbcT, GovT>where
WasmT: Wasm<<CustomT as Module>::ExecT, <CustomT as Module>::QueryT>,
BankT: Bank,
ApiT: Api,
StorageT: Storage,
CustomT: Module,
StakingT: Staking,
DistrT: Distribution,
IbcT: Ibc,
GovT: Gov,

pub fn init_modules<F, T>(&mut self, init_fn: F) -> Twhere
F: FnOnce(&mut Router<BankT, CustomT, WasmT, StakingT, DistrT, IbcT, GovT>, &dyn Api, &mut dyn Storage) -> T,

pub fn read_module<F, T>(&self, query_fn: F) -> Twhere
F: FnOnce(&Router<BankT, CustomT, WasmT, StakingT, DistrT, IbcT, GovT>, &dyn Api, &dyn Storage) -> T,

impl<BankT, ApiT, StorageT, CustomT, StakingT, DistrT, IbcT, GovT> App<BankT, ApiT, StorageT, CustomT, WasmKeeper<<CustomT as Module>::ExecT, <CustomT as Module>::QueryT>, StakingT, DistrT, IbcT, GovT>where
BankT: Bank,
ApiT: Api,
StorageT: Storage,
CustomT: Module,
StakingT: Staking,
DistrT: Distribution,
IbcT: Ibc,
GovT: Gov,
<CustomT as Module>::ExecT: 'static + Clone + Debug + PartialEq<<CustomT as Module>::ExecT> + JsonSchema + DeserializeOwned,
<CustomT as Module>::QueryT: 'static + CustomQuery + DeserializeOwned,

pub fn store_code(
&mut self,
code: Box<dyn Contract<<CustomT as Module>::ExecT, <CustomT as Module>::QueryT> + 'static, Global>
) -> u64

This registers contract code (like uploading wasm bytecode on a chain), so it can later be used to instantiate a contract.

pub fn contract_data(&self, address: &Addr) -> Result<ContractData, Error>

This allows to get ContractData for specific contract

pub fn dump_wasm_raw(
&self,
address: &Addr
) -> Vec<(Vec<u8, Global>, Vec<u8, Global>), Global>

This gets a raw state dump of all key-values held by a given contract

impl<BankT, ApiT, StorageT, CustomT, WasmT, StakingT, DistrT, IbcT, GovT> App<BankT, ApiT, StorageT, CustomT, WasmT, StakingT, DistrT, IbcT, GovT>where
<CustomT as Module>::ExecT: 'static + Debug + PartialEq<<CustomT as Module>::ExecT> + Clone + JsonSchema + DeserializeOwned,
<CustomT as Module>::QueryT: 'static + CustomQuery + DeserializeOwned,
WasmT: Wasm<<CustomT as Module>::ExecT, <CustomT as Module>::QueryT>,
BankT: Bank,
ApiT: Api,
StorageT: Storage,
CustomT: Module,
StakingT: Staking,
DistrT: Distribution,
IbcT: Ibc,
GovT: Gov,

pub fn set_block(&mut self, block: BlockInfo)

pub fn update_block<F>(&mut self, action: F)where
F: Fn(&mut BlockInfo),

pub fn block_info(&self) -> BlockInfo

Returns a copy of the current block_info

pub fn wrap(&self) -> QuerierWrapper<'_, <CustomT as Module>::QueryT>

Simple helper so we get access to all the QuerierWrapper helpers, eg. wrap().query_wasm_smart, query_all_balances, …

pub fn execute_multi(
&mut self,
sender: Addr,
msgs: Vec<CosmosMsg<<CustomT as Module>::ExecT>, Global>
) -> Result<Vec<AppResponse, Global>, Error>

Runs multiple CosmosMsg in one atomic operation. This will create a cache before the execution, so no state changes are persisted if any of them return an error. But all writes are persisted on success.

pub fn wasm_sudo<T, U>(
&mut self,
contract_addr: U,
msg: &T
) -> Result<AppResponse, Error>where
T: Serialize,
U: Into<Addr>,

Call a smart contract in “sudo” mode. This will create a cache before the execution, so no state changes are persisted if this returns an error, but all are persisted on success.

pub fn sudo(&mut self, msg: SudoMsg) -> Result<AppResponse, Error>

Runs arbitrary SudoMsg. This will create a cache before the execution, so no state changes are persisted if this returns an error, but all are persisted on success.