Struct vectis_contract_tests::common::common::ContractWrapper

pub struct ContractWrapper<T1, T2, T3, E1, E2, E3, C = Empty, Q = Empty, T4 = Empty, E4 = Error, E5 = Error, T6 = Empty, E6 = Error>where
    T1: DeserializeOwned + Debug,
    T2: DeserializeOwned,
    T3: DeserializeOwned,
    T4: DeserializeOwned,
    T6: DeserializeOwned,
    E1: 'static + Display + Debug + Send + Sync,
    E2: 'static + Display + Debug + Send + Sync,
    E3: 'static + Display + Debug + Send + Sync,
    E4: 'static + Display + Debug + Send + Sync,
    E5: 'static + Display + Debug + Send + Sync,
    E6: 'static + Display + Debug + Send + Sync,
    C: Clone + Debug + PartialEq<C> + JsonSchema,
    Q: 'static + CustomQuery + DeserializeOwned,{ /* private fields */ }

Wraps the exported functions from a contract and provides the normalized format Place T4 and E4 at the end, as we just want default placeholders for most contracts that don’t have sudo

Implementations

impl<T1, T2, T3, E1, E2, E3, C, Q> ContractWrapper<T1, T2, T3, E1, E2, E3, C, Q, Empty, Error, Error, Empty, Error>where
T1: 'static + DeserializeOwned + Debug,
T2: 'static + DeserializeOwned,
T3: 'static + DeserializeOwned,
E1: 'static + Display + Debug + Send + Sync,
E2: 'static + Display + Debug + Send + Sync,
E3: 'static + Display + Debug + Send + Sync,
C: 'static + Clone + Debug + PartialEq<C> + JsonSchema,
Q: 'static + CustomQuery + DeserializeOwned,

pub fn new(
execute_fn: fn(_: DepsMut<'_, Q>, _: Env, _: MessageInfo, _: T1) -> Result<Response<C>, E1>,
instantiate_fn: fn(_: DepsMut<'_, Q>, _: Env, _: MessageInfo, _: T2) -> Result<Response<C>, E2>,
query_fn: fn(_: Deps<'_, Q>, _: Env, _: T3) -> Result<Binary, E3>
) -> ContractWrapper<T1, T2, T3, E1, E2, E3, C, Q, Empty, Error, Error, Empty, Error>

pub fn new_with_empty(
execute_fn: fn(_: DepsMut<'_, Empty>, _: Env, _: MessageInfo, _: T1) -> Result<Response<Empty>, E1>,
instantiate_fn: fn(_: DepsMut<'_, Empty>, _: Env, _: MessageInfo, _: T2) -> Result<Response<Empty>, E2>,
query_fn: fn(_: Deps<'_, Empty>, _: Env, _: T3) -> Result<Binary, E3>
) -> ContractWrapper<T1, T2, T3, E1, E2, E3, C, Q, Empty, Error, Error, Empty, Error>

this will take a contract that returns Response and will “upgrade” it to Response if needed to be compatible with a chain-specific extension

impl<T1, T2, T3, E1, E2, E3, C, Q, T4, E4, E5, T6, E6> ContractWrapper<T1, T2, T3, E1, E2, E3, C, Q, T4, E4, E5, T6, E6>where
T1: 'static + DeserializeOwned + Debug,
T2: 'static + DeserializeOwned,
T3: 'static + DeserializeOwned,
T4: 'static + DeserializeOwned,
T6: 'static + DeserializeOwned,
E1: 'static + Display + Debug + Send + Sync,
E2: 'static + Display + Debug + Send + Sync,
E3: 'static + Display + Debug + Send + Sync,
E4: 'static + Display + Debug + Send + Sync,
E5: 'static + Display + Debug + Send + Sync,
E6: 'static + Display + Debug + Send + Sync,
C: 'static + Clone + Debug + PartialEq<C> + JsonSchema,
Q: 'static + CustomQuery + DeserializeOwned,

pub fn with_sudo<T4A, E4A>(
self,
sudo_fn: fn(_: DepsMut<'_, Q>, _: Env, _: T4A) -> Result<Response<C>, E4A>
) -> ContractWrapper<T1, T2, T3, E1, E2, E3, C, Q, T4A, E4A, E5, T6, E6>where
T4A: 'static + DeserializeOwned,
E4A: 'static + Display + Debug + Send + Sync,

pub fn with_sudo_empty<T4A, E4A>(
self,
sudo_fn: fn(_: DepsMut<'_, Q>, _: Env, _: T4A) -> Result<Response<Empty>, E4A>
) -> ContractWrapper<T1, T2, T3, E1, E2, E3, C, Q, T4A, E4A, E5, T6, E6>where
T4A: 'static + DeserializeOwned,
E4A: 'static + Display + Debug + Send + Sync,

pub fn with_reply<E5A>(
self,
reply_fn: fn(_: DepsMut<'_, Q>, _: Env, _: Reply) -> Result<Response<C>, E5A>
) -> ContractWrapper<T1, T2, T3, E1, E2, E3, C, Q, T4, E4, E5A, T6, E6>where
E5A: 'static + Display + Debug + Send + Sync,

pub fn with_reply_empty<E5A>(
self,
reply_fn: fn(_: DepsMut<'_, Q>, _: Env, _: Reply) -> Result<Response<Empty>, E5A>
) -> ContractWrapper<T1, T2, T3, E1, E2, E3, C, Q, T4, E4, E5A, T6, E6>where
E5A: 'static + Display + Debug + Send + Sync,

A correlate of new_with_empty

pub fn with_migrate<T6A, E6A>(
self,
migrate_fn: fn(_: DepsMut<'_, Q>, _: Env, _: T6A) -> Result<Response<C>, E6A>
) -> ContractWrapper<T1, T2, T3, E1, E2, E3, C, Q, T4, E4, E5, T6A, E6A>where
T6A: 'static + DeserializeOwned,
E6A: 'static + Display + Debug + Send + Sync,

pub fn with_migrate_empty<T6A, E6A>(
self,
migrate_fn: fn(_: DepsMut<'_, Q>, _: Env, _: T6A) -> Result<Response<Empty>, E6A>
) -> ContractWrapper<T1, T2, T3, E1, E2, E3, C, Q, T4, E4, E5, T6A, E6A>where
T6A: 'static + DeserializeOwned,
E6A: 'static + Display + Debug + Send + Sync,