Struct vectis_contract_tests::common::common::Message

pub struct Message(_);

A (hashed) message input to an ECDSA signature

Implementations

impl Message

pub fn as_ptr(&self) -> *const u8

Converts the object to a raw pointer for FFI interfacing

pub fn as_mut_ptr(&mut self) -> *mut u8

Converts the object to a mutable raw pointer for FFI interfacing

pub fn len(&self) -> usize

Returns the length of the object as an array

pub fn is_empty(&self) -> bool

Returns whether the object as an array is empty

impl Message

pub fn from_slice(data: &[u8]) -> Result<Message, Error>

If you just want to sign an arbitrary message use Message::from_hashed_data instead.

Converts a MESSAGE_SIZE-byte slice to a message object. WARNING: the slice has to be a cryptographically secure hash of the actual message that’s going to be signed. Otherwise the result of signing isn’t a secure signature.

pub fn from_hashed_data<H>(data: &[u8]) -> Messagewhere H: ThirtyTwoByteHash + Hash,

Constructs a Message by hashing data with hash algorithm H. This requires the feature bitcoin_hashes to be enabled.

extern crate bitcoin_hashes;
use secp256k1::Message;
use bitcoin_hashes::sha256;
use bitcoin_hashes::Hash;

let m1 = Message::from_hashed_data::<sha256::Hash>("Hello world!".as_bytes());
// is equivalent to
let m2 = Message::from(sha256::Hash::hash("Hello world!".as_bytes()));

assert_eq!(m1, m2);