Struct vectis_contract_tests::common::common::Secp256k1

pub struct Secp256k1<C>where
    C: Context,{ /* private fields */ }

The secp256k1 engine, used to execute all signature operations

Implementations

impl<C> Secp256k1<C>where C: Context,

pub fn gen_new() -> Secp256k1<C>

Lets you create a context in a generic manner(sign/verify/all)

impl Secp256k1<All>

pub fn new() -> Secp256k1<All>

Creates a new Secp256k1 context with all capabilities

impl Secp256k1<SignOnly>

pub fn signing_only() -> Secp256k1<SignOnly>

Creates a new Secp256k1 context that can only be used for signing

impl Secp256k1<VerifyOnly>

pub fn verification_only() -> Secp256k1<VerifyOnly>

Creates a new Secp256k1 context that can only be used for verification

impl<'buf, C> Secp256k1<C>where C: Context + 'buf,

pub fn preallocated_gen_new( buf: &'buf mut [AlignedType] ) -> Result<Secp256k1<C>, Error>

Lets you create a context with preallocated buffer in a generic manner(sign/verify/all)

impl<'buf> Secp256k1<AllPreallocated<'buf>>

pub fn preallocated_new( buf: &'buf mut [AlignedType] ) -> Result<Secp256k1<AllPreallocated<'buf>>, Error>

Creates a new Secp256k1 context with all capabilities

pub fn preallocate_size() -> usize

Uses the ffi secp256k1_context_preallocated_size to check the memory size needed for a context

pub unsafe fn from_raw_all( raw_ctx: *mut Context ) -> ManuallyDrop<Secp256k1<AllPreallocated<'buf>>>

Create a context from a raw context.

Safety

This is highly unsafe, due to the number of conditions that aren’t checked.

raw_ctx needs to be a valid Secp256k1 context pointer. that was generated by exactly the same code/version of the libsecp256k1 used here.

The capabilities (All/SignOnly/VerifyOnly) of the context must match the flags passed to libsecp256k1 when generating the context.

The user must handle the freeing of the context(using the correct functions) by himself.

Violating these may lead to Undefined Behavior.

impl<'buf> Secp256k1<SignOnlyPreallocated<'buf>>

pub fn preallocated_signing_only( buf: &'buf mut [AlignedType] ) -> Result<Secp256k1<SignOnlyPreallocated<'buf>>, Error>

Creates a new Secp256k1 context that can only be used for signing

pub fn preallocate_signing_size() -> usize

Uses the ffi secp256k1_context_preallocated_size to check the memory size needed for the context

pub unsafe fn from_raw_signining_only( raw_ctx: *mut Context ) -> ManuallyDrop<Secp256k1<SignOnlyPreallocated<'buf>>>

Create a context from a raw context.

Safety

This is highly unsafe, due to the number of conditions that aren’t checked.

raw_ctx needs to be a valid Secp256k1 context pointer. that was generated by exactly the same code/version of the libsecp256k1 used here.

The capabilities (All/SignOnly/VerifyOnly) of the context must match the flags passed to libsecp256k1 when generating the context.

The user must handle the freeing of the context(using the correct functions) by himself.

This list is not exhaustive, and any violation may lead to Undefined Behavior.,

impl<'buf> Secp256k1<VerifyOnlyPreallocated<'buf>>

pub fn preallocated_verification_only( buf: &'buf mut [AlignedType] ) -> Result<Secp256k1<VerifyOnlyPreallocated<'buf>>, Error>

Creates a new Secp256k1 context that can only be used for verification

pub fn preallocate_verification_size() -> usize

Uses the ffi secp256k1_context_preallocated_size to check the memory size needed for the context

pub unsafe fn from_raw_verification_only( raw_ctx: *mut Context ) -> ManuallyDrop<Secp256k1<VerifyOnlyPreallocated<'buf>>>

Create a context from a raw context.

Safety

This is highly unsafe, due to the number of conditions that aren’t checked.

raw_ctx needs to be a valid Secp256k1 context pointer. that was generated by exactly the same code/version of the libsecp256k1 used here.

The capabilities (All/SignOnly/VerifyOnly) of the context must match the flags passed to libsecp256k1 when generating the context.

The user must handle the freeing of the context(using the correct functions) by himself.

This list is not exhaustive, and any violation may lead to Undefined Behavior.,

impl<C> Secp256k1<C>where C: Signing,

pub fn schnorrsig_sign_no_aux_rand( &self, msg: &Message, keypair: &KeyPair ) -> Signature

Create a schnorr signature without using any auxiliary random data.

pub fn schnorrsig_sign_with_aux_rand( &self, msg: &Message, keypair: &KeyPair, aux_rand: &[u8; 32] ) -> Signature

Create a Schnorr signature using the given auxiliary random data.

pub fn schnorrsig_verify( &self, sig: &Signature, msg: &Message, pubkey: &PublicKey ) -> Result<(), Error>

Verify a Schnorr signature.

impl<C> Secp256k1<C>where C: Context,

pub fn ctx(&self) -> &*mut Context

Getter for the raw pointer to the underlying secp256k1 context. This shouldn’t be needed with normal usage of the library. It enables extending the Secp256k1 with more cryptographic algorithms outside of this crate.

pub fn preallocate_size_gen() -> usize

Returns the required memory for a preallocated context buffer in a generic manner(sign/verify/all)

pub fn seeded_randomize(&mut self, seed: &[u8; 32])

(Re)randomizes the Secp256k1 context for cheap sidechannel resistance given 32 bytes of cryptographically-secure random data; see comment in libsecp256k1 commit d2275795f by Gregory Maxwell.

impl<C> Secp256k1<C>where C: Signing,

pub fn sign(&self, msg: &Message, sk: &SecretKey) -> Signature

Constructs a signature for msg using the secret key sk and RFC6979 nonce Requires a signing-capable context.

pub fn sign_grind_r( &self, msg: &Message, sk: &SecretKey, bytes_to_grind: usize ) -> Signature

Constructs a signature for msg using the secret key sk, RFC6979 nonce and “grinds” the nonce by passing extra entropy if necessary to produce a signature that is less than 71 - bytes_to_grund bytes. The number of signing operation performed by this function is exponential in the number of bytes grinded. Requires a signing capable context.

pub fn sign_low_r(&self, msg: &Message, sk: &SecretKey) -> Signature

Constructs a signature for msg using the secret key sk, RFC6979 nonce and “grinds” the nonce by passing extra entropy if necessary to produce a signature that is less than 71 bytes and compatible with the low r signature implementation of bitcoin core. In average, this function will perform two signing operations. Requires a signing capable context.

impl<C> Secp256k1<C>where C: Verification,

pub fn verify( &self, msg: &Message, sig: &Signature, pk: &PublicKey ) -> Result<(), Error>

Checks that sig is a valid ECDSA signature for msg using the public key pubkey. Returns Ok(()) on success. Note that this function cannot be used for Bitcoin consensus checking since there may exist signatures which OpenSSL would verify but not libsecp256k1, or vice-versa. Requires a verify-capable context.

let message = Message::from_slice(&[0xab; 32]).expect("32 bytes");
let sig = secp.sign(&message, &secret_key);
assert_eq!(secp.verify(&message, &sig, &public_key), Ok(()));

let message = Message::from_slice(&[0xcd; 32]).expect("32 bytes");
assert_eq!(secp.verify(&message, &sig, &public_key), Err(Error::IncorrectSignature));