Struct vectis_contract_tests::common::remote_common::RemoteChainSuite

pub struct RemoteChainSuite {
    pub app: App,
    pub controller: Addr,
    pub factory: Addr,
    pub remote_tunnel: Addr,
}

Fields
app: Appcontroller: Addr

Admin Addr (for test we use admin, but it should be self)

factory: Addrremote_tunnel: Addr
Implementations

impl RemoteChainSuite

pub fn init() -> Result<RemoteChainSuite>

In reality we instantiate remote_runnel contract Use remote_tunnel to instantiate_factory from DAO on dao-chain

But we cannot access the ibc hooks so we do the factory first so that we can fill in the factory addr

pub fn create_new_proxy_without_guardians(
&mut self,
controller: Addr,
factory: Addr,
proxy_initial_fund: Vec<Coin>,
guardians_multisig: Option<MultiSig>,
native_tokens_amount: u128
) -> Result<Addr>

pub fn create_new_proxy(
&mut self,
controller: Addr,
proxy_initial_funds: Vec<Coin>,
guardians_multisig: Option<MultiSig>,
native_tokens_amount: u128
) -> Result<Addr>

pub fn create_relay_transaction(
&mut self,
signer_sk: &[u8; 32],
cosmos_msg: CosmosMsg,
nonce: u64
) -> RelayTransaction

pub fn update_proxy_code_id(
&mut self,
new_code_id: u64,
factory: Addr
) -> Result<AppResponse>

pub fn update_proxy_multisig_code_id(
&mut self,
new_code_id: u64,
factory: Addr
) -> Result<AppResponse>

pub fn create_guardians_request_and_update_guardians(
&mut self,
sender: &Addr,
wallet_address: &Addr,
guardians: Guardians,
new_multisig_code_id: Option<u64>
) -> Result<AppResponse>

pub fn fast_forward_block_time(&mut self, forward_time_sec: u64)

pub fn query_unclaimed_govec_wallets(
&self,
contract_addr: &Addr,
start_after: Option<String>,
limit: Option<u32>
) -> Result<UnclaimedWalletList, StdError>

pub fn query_proxy_govec_claim_expiration(
&self,
contract_addr: &Addr,
proxy: &Addr
) -> Result<Option<Expiration>, StdError>

pub fn query_wallet_info<R>(&self, contract_addr: &Addr) -> Result<R, StdError>where
R: DeserializeOwned,

pub fn query_guardians_request(
&self,
contract_addr: &Addr
) -> Result<Option<GuardiansUpdateRequest>, StdError>

pub fn query_multisig_voters(
&self,
contract_addr: &Addr
) -> Result<VoterListResponse, StdError>

pub fn query_balance(&self, addr: &Addr) -> Result<Coin>

pub fn claim_expiration(&self) -> u64