Trait vectis_contract_tests::common::common::Executor

pub trait Executor<C>where
    C: Clone + Debug + PartialEq<C> + JsonSchema + 'static,{
    fn execute(
        &mut self,
        sender: Addr,
        msg: CosmosMsg<C>
    ) -> Result<AppResponse, Error>;

    fn instantiate_contract<T, U>(
        &mut self,
        code_id: u64,
        sender: Addr,
        init_msg: &T,
        send_funds: &[Coin],
        label: U,
        admin: Option<String>
    ) -> Result<Addr, Error>
    where
        T: Serialize,
        U: Into<String>,
    { ... }
    fn execute_contract<T>(
        &mut self,
        sender: Addr,
        contract_addr: Addr,
        msg: &T,
        send_funds: &[Coin]
    ) -> Result<AppResponse, Error>
    where
        T: Serialize + Debug,
    { ... }
    fn migrate_contract<T>(
        &mut self,
        sender: Addr,
        contract_addr: Addr,
        msg: &T,
        new_code_id: u64
    ) -> Result<AppResponse, Error>
    where
        T: Serialize,
    { ... }
    fn send_tokens(
        &mut self,
        sender: Addr,
        recipient: Addr,
        amount: &[Coin]
    ) -> Result<AppResponse, Error> { ... }
}

Required Methods

fn execute(
&mut self,
sender: Addr,
msg: CosmosMsg<C>
) -> Result<AppResponse, Error>

Runs arbitrary CosmosMsg. This will create a cache before the execution, so no state changes are persisted if this returns an error, but all are persisted on success.

Provided Methods

fn instantiate_contract<T, U>(
&mut self,
code_id: u64,
sender: Addr,
init_msg: &T,
send_funds: &[Coin],
label: U,
admin: Option<String>
) -> Result<Addr, Error>where
T: Serialize,
U: Into<String>,

Create a contract and get the new address. This is just a helper around execute()

fn execute_contract<T>(
&mut self,
sender: Addr,
contract_addr: Addr,
msg: &T,
send_funds: &[Coin]
) -> Result<AppResponse, Error>where
T: Serialize + Debug,

Execute a contract and process all returned messages. This is just a helper around execute(), but we parse out the data field to that what is returned by the contract (not the protobuf wrapper)

fn migrate_contract<T>(
&mut self,
sender: Addr,
contract_addr: Addr,
msg: &T,
new_code_id: u64
) -> Result<AppResponse, Error>where
T: Serialize,

Migrate a contract. Sender must be registered admin. This is just a helper around execute()

fn send_tokens(
&mut self,
sender: Addr,
recipient: Addr,
amount: &[Coin]
) -> Result<AppResponse, Error>